+++ src/cpp/engine/collision/bounding_box.h
@@ -12,21 +12,21 @@ namespace engine {
 
 class BoundingBox {
  protected:
-  glm::vec3 mins_;
-  glm::vec3 maxes_;
+  glm::dvec3 mins_;
+  glm::dvec3 maxes_;
  public:
   BoundingBox() = default;
 
-  BoundingBox(const glm::vec3& mins, const glm::vec3& maxes)
+  BoundingBox(const glm::dvec3& mins, const glm::dvec3& maxes)
       : mins_(mins), maxes_(maxes) {}
 
-  glm::vec3 mins() const { return mins_; }
-  glm::vec3 maxes() const { return maxes_; }
-  glm::vec3 center() const { return (maxes_+mins_) / 2.0f; }
-  glm::vec3 extent() const { return maxes_-mins_; }
+  glm::dvec3 mins() const { return mins_; }
+  glm::dvec3 maxes() const { return maxes_; }
+  glm::dvec3 center() const { return (maxes_+mins_) / 2.0; }
+  glm::dvec3 extent() const { return maxes_-mins_; }
 
-  virtual bool collidesWithSphere(const glm::vec3& center, float radius) const {
-    float dmin = 0;
+  virtual bool collidesWithSphere(const glm::dvec3& center, double radius) const {
+    double dmin = 0;
     for (int i = 0; i < 3; ++i) {
       if (center[i] < mins_[i]) {
         dmin += sqr(center[i] - mins_[i]);
@@ -38,14 +38,14 @@ class BoundingBox {
   }
 
   virtual bool collidesWithFrustum(const Frustum& frustum) const {
-    glm::vec3 center = this->center();
-    glm::vec3 extent = this->extent();
+    glm::dvec3 center = this->center();
+    glm::dvec3 extent = this->extent();
 
     for(int i = 0; i < 6; ++i) {
       const Plane& plane = frustum.planes[i];
 
-      float d = glm::dot(center, plane.normal);
-      float r = glm::dot(extent, glm::abs(plane.normal));
+      double d = glm::dot(center, plane.normal);
+      double r = glm::dot(extent, glm::abs(plane.normal));
 
       if(d + r < -plane.dist) {
         return false;
